Extract the crop and tool surface classes into their own file
`CropView.swift` was 3463 lines, ~670 of which were the surface classes:
`ViewportRenderingSurface`, `ViewportRenderingState`, `SurfaceHost`,
`CanvasSurface`, `CropSurface`, and `ToolSurface`. They are self-contained —
they never reach into `CropView` state except through the weak `owner`
back-pointer and the stored event closures — and they own the scroll,
canvas, and bounce-ride mechanics that evolve together.

Move them verbatim into `CropView.CanvasSurface.swift` as
`extension CropView { ... }`, mirroring the existing
`CropView._CropScrollView.swift` pattern, so every call site keeps its
`CropView`-nested spelling and compiles unchanged.

This is a pure move: the relocated lines are byte-identical to the removed
ones except for the access-control widenings listed below. No logic, no
renames, no reordering. The display-link state machine and the geometry
builders stay in CropView.swift.

Access-control changes forced by the move (a `private`/`fileprivate` member
is not visible from an extension in another file):

- the six moved type declarations: `private` -> internal
- `CropView.CanvasStrokeCommitHandler`: `private` -> internal
- `CropView.CanvasInputKey`: `private` -> internal
- `CropView.viewportRenderingDisplayLinkDidTick(_:surface:)`: `private` ->
  internal
- `UIEdgeInsets.multiplied(_:)`: `fileprivate` -> internal
- `CGRect.isNearlyEqual(to:tolerance:)`: `fileprivate` -> internal
- `CropDisplayViewport.editingCanvasViewport`: `fileprivate` -> internal
